backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
yfinance==0.2.38
pandas==2.2.2
numpy==1.26.4
transformers==4.41.0
torch==2.3.0
requests==2.32.3
python-dotenv==1.0.1
matplotlib==3.9.0
plotly==5.22.0
